factor fetchPost into common.js
